WIP Add missing file configuration functionality
On https://github.com/OCA/connector/pull/210 I added the possibility
to configure the runner channels thru the odoo configuration file.

This was ported to v9 on PR 113, but not to 10 yet...

Which likely explains why queue_job doesn't contain this functionality

This empty commit is a placeholder for a WIP PR for porting that
functionality to the queue module.